Rename syntaxexamples 'plugins' to 'plugin'
