+++ api-test/integration-tests/src/test/java/com/microsoft/semantickernel/tests/WiremockExamplesIT.java
@@ -32,8 +32,8 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.Queue;
 import java.util.stream.Stream;
-import com.microsoft.semantickernel.samples.syntaxexamples.plugins.Example10_DescribeAllPluginsAndFunctions;
-import com.microsoft.semantickernel.samples.syntaxexamples.plugins.Example13_ConversationSummaryPlugin;
+import com.microsoft.semantickernel.samples.syntaxexamples.plugin.Example10_DescribeAllPluginsAndFunctions;
+import com.microsoft.semantickernel.samples.syntaxexamples.plugin.Example13_ConversationSummaryPlugin;
 import com.microsoft.semantickernel.samples.syntaxexamples.template.Example06_TemplateLanguage;
 import com.microsoft.semantickernel.samples.syntaxexamples.template.Example56_TemplateMethodFunctionsWithMultipleArguments;
 import com.microsoft.semantickernel.samples.syntaxexamples.template.Example64_MultiplePromptTemplates;
